Weight/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
Werkzeug==2.3.7
pytest==7.4.4
requests

